refactor(backend): colocate evasion domain into feature package
Moves the evasion controller, entity, repository and DTO out of the
generic controller/database/dto layer packages into
net.onelitefeather.blackhole.backend.evasion, alongside the
already-colocated IpCorrelationService/IpCorrelationRetentionSweeper.
Fully self-contained domain, no external callers to update. Pure
physical move, no behavior change.
